move gvk for CAPI Cluster to common package
