+++ cluster-operator/controllers/capi/capi_cluster_controller.go
@@ -6,6 +6,8 @@ package capi
 import (
 	"context"
 	"fmt"
+
+	"github.com/verrazzano/verrazzano/cluster-operator/internal/capi"
 	"github.com/verrazzano/verrazzano/pkg/constants"
 	vzstring "github.com/verrazzano/verrazzano/pkg/string"
 	"go.uber.org/zap"
@@ -17,7 +19,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
@@ -45,15 +46,9 @@ type CAPIClusterReconciler struct {
 	VerrazzanoRegistrar *VerrazzanoRegistration
 }
 
-var gvk = schema.GroupVersionKind{
-	Group:   "cluster.x-k8s.io",
-	Version: "v1beta1",
-	Kind:    "Cluster",
-}
-
 func CAPIClusterClientObject() client.Object {
 	obj := &unstructured.Unstructured{}
-	obj.SetGroupVersionKind(gvk)
+	obj.SetGroupVersionKind(capi.GVKCAPICluster)
 	return obj
 }
 
@@ -69,7 +64,7 @@ func (r *CAPIClusterReconciler) Reconcile(ctx context.Context, req ctrl.Request)
 	r.Log.Infof("Reconciling CAPI cluster: %v", req.NamespacedName)
 
 	cluster := &unstructured.Unstructured{}
-	cluster.SetGroupVersionKind(gvk)
+	cluster.SetGroupVersionKind(capi.GVKCAPICluster)
 	err := r.Get(context.TODO(), req.NamespacedName, cluster)
 	if err != nil && !errors.IsNotFound(err) {
 		return ctrl.Result{}, err
